+++ app/src/main/java/com/tanishkej/onerep/data/datasource/WorkoutDataSource.kt
@@ -12,7 +12,7 @@ import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOn
 import java.text.DateFormat
 import java.text.SimpleDateFormat
-import java.util.*
+import java.util.Date
 import java.util.Locale
 import javax.inject.Inject
