+++ app/src/main/java/com/tanishkej/onerep/data/repository/WorkoutRepository.kt
@@ -1,6 +1,5 @@
 package com.tanishkej.onerep.data.repository
 
-import com.tanishkej.onerep.data.model.Workout
 import com.tanishkej.onerep.data.model.WorkoutGroups
 import kotlinx.coroutines.flow.Flow
