+++ app/src/main/java/com/tanishkej/onerep/ui/workouts/screens/WorkoutScreen.kt
@@ -6,35 +6,26 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.tanishkej.onerep.ui.workouts.states.WorkoutUiState
-import androidx.compose.material3.Text
 import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
-import com.github.mikephil.charting.charts.LineChart
-import com.github.mikephil.charting.data.Entry
-import com.github.mikephil.charting.data.LineData
-import com.github.mikephil.charting.data.LineDataSet
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.tanishkej.onerep.R
-import com.tanishkej.onerep.data.model.Workout
 import com.tanishkej.onerep.data.util.WorkoutUtil.getGraphEntries
-import com.tanishkej.onerep.ui.components.OneRepGraph
 import com.tanishkej.onerep.ui.components.OneRepLoadingWheel
+import com.tanishkej.onerep.ui.components.graphs.OneRepGraph
 import com.tanishkej.onerep.ui.workouts.cards.WorkoutCard
-import com.tanishkej.onerep.ui.workouts.states.WorkoutListUiState
+import com.tanishkej.onerep.ui.workouts.states.WorkoutUiState
 import com.tanishkej.onerep.ui.workouts.viewmodels.WorkoutViewModel
 
 @Composable
